+++ src/main/java/io/swagger/dao/RepositoryUser.java
@@ -18,7 +18,7 @@ public interface RepositoryUser extends CrudRepository<User, Long> {
     @Transactional
     @Modifying
     @Query("UPDATE User u SET u.firstname = :firstname, u.lastname = :lastname, u.email = :email, u.phone = :phone, u.birthdate = :birthdate, u.rank = :rank, u.status = :status  WHERE u.id = :userId")
-    void updateUser(@Param("userId") Integer userId, @Param("firstname") String firstname, @Param("lastname") String lastname, @Param("email") String email, @Param("phone") String phone, @Param("birthdate") String birthdate, @Param("rank") User.RankEnum rank, @Param("status") User.StatusEnum status);
+    void updateUser(@Param("userId") Long userId, @Param("firstname") String firstname, @Param("lastname") String lastname, @Param("email") String email, @Param("phone") String phone, @Param("birthdate") String birthdate, @Param("rank") User.RankEnum rank, @Param("status") User.StatusEnum status);
 
     @Query("select u from User u where u.email =:email and u.password = :password")
     User findUserByUserCredentials(@Param("email") String username, @Param("password") String password);
